Update SeventhSeaRoll to allow for toggling between exploding and non-exploding states
Either an individual die, or a complete roll can be toggled between exploding and not.
If the roll is toggle, then all dice within the roll are also toggled. However, when
toggling a die, the die will still be rolled as if it were exploding, though when
queried for the facing number, it will only return 10.